audio and visual merged, data collection ready for device debug
